requirements.txt
# Alfred Gemini - Dependencies

# Google AI (primary)
google-generativeai>=0.3.0

# Audio I/O
sounddevice>=0.4.6
numpy>=1.26.4

# Fallback speech recognition
SpeechRecognition>=3.10.0

# Environment
python-dotenv>=1.0.1

# Optional: System TTS fallback (Windows)
# pyttsx3>=2.90

# Development
pytest>=8.0.0
